+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+* Replaced the balancer connection to discovery service from short-lived grpc connection to `internal/conn` lazy connection (revert related changes from `v3.42.6`)
+* Marked as deprecated `trace.Driver.OnBalancerDialEntrypoint` event callback
+* Deprecated `trace.Driver.OnConnTake` event callback
+* Added `trace.Driver.OnConnDial` event callback
+
 ## v3.42.11
 * Fixed validation error for topicoptions.WithPartitionID option of start topic writer.